Change how database tests are performed
Begin transitioning to pytest's style of unit testing. We do this so
that it is easier to write tests especially by taking advantage of
pytest's reusable fixtures.

Running new database tests with pytest allows us to easily load a
database in memory so that we don't modify a database file on each
test execution. This means that we don't have to contend with it when
managing changes in git and we don't have to delete modifications from
old tests whenever we want to run a new test.