Return last activity chronologically
Fix bug where when we record an activity at a time before one that has
already recorded, the returned last activity is the one that was
recorded most recently, instead of the activity that occurred just
before the newly recorded activity. This resulted in showing the user
the wrong last activity along with a negative duration of time.

Now the user is shown the activity that took place directly before the
activity they just recorded.